checking for existence of index first now

+++ colplus-dao/src/main/java/org/col/es/NameUsageIndexService.java
@@ -59,10 +59,12 @@ public class NameUsageIndexService {
   public void indexDataset(final int datasetKey) {
     final String index = NAME_USAGE_BASE;
     final int batchSize = esConfig.nameUsage.batchSize;
-    EsUtil.deleteDataset(client, index, datasetKey);
-    // Probably nor necessary, but let's do it anyhow.
-    EsUtil.refreshIndex(client, index);
-    EsUtil.createIndex(client, index, esConfig.nameUsage);
+    if (EsUtil.indexExists(client, index)) {
+      EsUtil.deleteDataset(client, index, datasetKey);
+      EsUtil.refreshIndex(client, index);
+    } else {
+      EsUtil.createIndex(client, index, esConfig.nameUsage);
+    }
     final AtomicInteger counter = new AtomicInteger();
     try (SqlSession session = factory.openSession()) {
       NameUsageMapper mapper = session.getMapper(NameUsageMapper.class);
